remove Contact entity file and integrate relevant methods into Request enttity file
