+++ app/entities/Request.py
@@ -3,7 +3,6 @@ from uuid import UUID, uuid4
 from datetime import datetime
 from sqlmodel import SQLModel, Field, Relationship
 from ..entities.GroundStation import GroundStation
-from ..entities.Contact import Contact
 from pydantic import ConfigDict
 
 if TYPE_CHECKING:
@@ -31,6 +30,33 @@ class RFRequest(SQLModel, table=True):
     ground_station_id: Optional[int] = Field(
         default=None, foreign_key="ground_stations.id"
     )
+    time_remaining: int
+    num_passes_remaining: int
+
+    # copying over from old RFTime class
+    def get_priority_weight(self) -> float:
+        tot_time = (
+            self.uplink_time_requested
+            + self.downlink_time_requested
+            + self.science_time_requested
+        )
+        time_period = self.end_time - self.start_time
+
+        return (self.end_time - datetime.now()).total_seconds() * (
+            tot_time / time_period.total_seconds()
+        )
+
+    def set_time_remaining(self, time_booked: int):
+        self.timeRemaining = self.time_remaining - time_booked
+
+    def decrease_pass(self):
+        self.num_passes_remaining -= 1
+
+    def __repr__(self):
+        return (
+            f"RFTime(gs={self.ground_station_id}, sat={self.satellite_id}, "
+            f"start={self.start_time}, end={self.end_time}, dur={(self.end_time-self.start_time).total_seconds()}s)"
+        )
 
     # satellite: Optional["Satellite"] = Relationship(back_populates="requests")
     # contact: Optional["Contact"] = Relationship()
@@ -67,6 +93,15 @@ class ContactRequest(SQLModel, table=True):
     # satellite: Optional["Satellite"] = Relationship(back_populates="requests")
     # contact: Optional["Contact"] = Relationship()
     # ground_station: Optional["GroundStation"] = Relationship()
+
+    def __repr__(self):
+        duration_str = "N/A"
+        if self.aos and self.los:
+            duration_str = f"{(self.los-self.aos).total_seconds()}s"
+        return (
+            f"Contact(gs={self.ground_station_id}, sat={self.satellite_id}, "
+            f"start={self.aos}, end={self.los}, dur={duration_str})"
+        )
 
 
 # RequestBase = Union[RFRequest, ContactRequest]
